(svn r11170) -Fix (r11152): a variable's type was too small for the data that is going to put into it when NewIndustries are finally activated. For now it just removes a warning and doesn't really fix anything.

+++ src/industry.h
@@ -9,7 +9,7 @@
 #include "helpers.hpp"
 #include "newgrf_storage.h"
 
-typedef byte IndustryGfx;
+typedef uint16 IndustryGfx;
 typedef uint8 IndustryType;
 
 enum {
